First Version of HumanAI
Added the Human/Soldier Enemy AI script. This will control the movement, attack, shield (if allowed), and Phase two (if applied). The script will work for the BasicSoldierEnemy, GeneralSoldierEnemy, and the JuggernautEnemy. This code is mostly done for its first pass. It just needs a few things added and checked. After these small things are done, the soldier enemies need to be correctly linked to their models, materials, and animations.