Update Working Dungeon Scene May 2 8:40
Uploaded the working GeneralSodierAI, it only has an Idle and an Attack animation. The prefab is HumanEnemy in the Prefab folder (not in the GeneralEnemyParts folder (which is all of the stuff for the enemy). I added the sound audio sources. The Human Enemy sound is completely set up. The Mage sound is attached to the Mage enemy and set to make a sound when the enemy attacks. The Charge Enemy and Vampire sounds are in the dungeon Scene but I don't have their enemies to hook them up to. The Range Enemy sounds are in the Dungeon Scene but need to be set up for the animation. The Player-based sounds are also in the scene, but I do not know where the triggers are in the Player scripts. The sound scripts often lose their values and must be set when added to the scenes.